refactor(AnswerViewPage): 기존 FeedPage를 ViewPage로 분리 및 코드 정리
