+++ src/pages/FeedPage/AnswerViewPage.jsx
@@ -1,9 +1,11 @@
 import AnswerCluster from './components/AnswerCluster';
+import FloatingButton from './components/FloatingButton';
 
 function AnswerViewPage() {
   return (
     <>
       <AnswerCluster isEditable={false} />
+      <FloatingButton />
     </>
   );
 }
