+++ src/pages/FeedPage/components/FloatingButton.jsx
@@ -5,21 +5,14 @@ import styled from 'styled-components';
 
 const viewTxt = window.innerWidth < 768 ? '질문 작성' : '질문 작성하기';
 
-const FloatingButton = ({ setQuestions, name, imageSource }) => {
+const FloatingButton = () => {
   const [isOpen, setIsOpen] = useState(false);
   return (
     <>
       <StyledFloatingButton onClick={() => setIsOpen(true)}>
         {viewTxt}
       </StyledFloatingButton>
-      {isOpen && (
-        <Modal
-          setQuestions={setQuestions}
-          name={name}
-          imageSource={imageSource}
-          onClose={() => setIsOpen(false)}
-        />
-      )}
+      {isOpen && <Modal onClose={() => setIsOpen(false)} />}
     </>
   );
 };
